static morph. operations to their own class
